Recreated demo project.
