+++ DatePickerDialog.xcodeproj/xcshareddata/xcschemes/DatePickerDialog.xcscheme
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Scheme
-   LastUpgradeVersion = "0700"
+   LastUpgradeVersion = "0800"
    version = "1.3">
    <BuildAction
       parallelizeBuildables = "YES"
@@ -14,10 +14,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "81EC6C7B1A1EADF7009440DB"
-               BuildableName = "DatePickerDialogExample.app"
-               BlueprintName = "DatePickerDialogExample"
-               ReferencedContainer = "container:DatePickerDialogExample.xcodeproj">
+               BlueprintIdentifier = "4BDE05A21D8F3AF000DEF7F3"
+               BuildableName = "DatePickerDialog.framework"
+               BlueprintName = "DatePickerDialog"
+               ReferencedContainer = "container:DatePickerDialog.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -28,14 +28,24 @@
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       shouldUseLaunchSchemeArgsEnv = "YES">
       <Testables>
+         <TestableReference
+            skipped = "NO">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "4BDE05AB1D8F3AF000DEF7F3"
+               BuildableName = "DatePickerDialogTests.xctest"
+               BlueprintName = "DatePickerDialogTests"
+               ReferencedContainer = "container:DatePickerDialog.xcodeproj">
+            </BuildableReference>
+         </TestableReference>
       </Testables>
       <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "81EC6C7B1A1EADF7009440DB"
-            BuildableName = "DatePickerDialogExample.app"
-            BlueprintName = "DatePickerDialogExample"
-            ReferencedContainer = "container:DatePickerDialogExample.xcodeproj">
+            BlueprintIdentifier = "4BDE05A21D8F3AF000DEF7F3"
+            BuildableName = "DatePickerDialog.framework"
+            BlueprintName = "DatePickerDialog"
+            ReferencedContainer = "container:DatePickerDialog.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -51,16 +61,15 @@
       debugDocumentVersioning = "YES"
       debugServiceExtension = "internal"
       allowLocationSimulation = "YES">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
+      <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "81EC6C7B1A1EADF7009440DB"
-            BuildableName = "DatePickerDialogExample.app"
-            BlueprintName = "DatePickerDialogExample"
-            ReferencedContainer = "container:DatePickerDialogExample.xcodeproj">
+            BlueprintIdentifier = "4BDE05A21D8F3AF000DEF7F3"
+            BuildableName = "DatePickerDialog.framework"
+            BlueprintName = "DatePickerDialog"
+            ReferencedContainer = "container:DatePickerDialog.xcodeproj">
          </BuildableReference>
-      </BuildableProductRunnable>
+      </MacroExpansion>
       <AdditionalOptions>
       </AdditionalOptions>
    </LaunchAction>
@@ -70,16 +79,15 @@
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
       debugDocumentVersioning = "YES">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
+      <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "81EC6C7B1A1EADF7009440DB"
-            BuildableName = "DatePickerDialogExample.app"
-            BlueprintName = "DatePickerDialogExample"
-            ReferencedContainer = "container:DatePickerDialogExample.xcodeproj">
+            BlueprintIdentifier = "4BDE05A21D8F3AF000DEF7F3"
+            BuildableName = "DatePickerDialog.framework"
+            BlueprintName = "DatePickerDialog"
+            ReferencedContainer = "container:DatePickerDialog.xcodeproj">
          </BuildableReference>
-      </BuildableProductRunnable>
+      </MacroExpansion>
    </ProfileAction>
    <AnalyzeAction
       buildConfiguration = "Debug">
