+++ DatePickerDialog.xcodeproj/xcshareddata/xcschemes/DatePickerDialogExample.xcscheme
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Scheme
-   LastUpgradeVersion = "0730"
+   LastUpgradeVersion = "0800"
    version = "1.3">
    <BuildAction
       parallelizeBuildables = "YES"
@@ -14,10 +14,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "BC24AA5A1CDB5D7A006438C4"
-               BuildableName = "DatePickerDialogFramework.framework"
-               BlueprintName = "DatePickerDialogFramework"
-               ReferencedContainer = "container:DatePickerDialogExample.xcodeproj">
+               BlueprintIdentifier = "4B1E9F361D8F3C9A00C1EF25"
+               BuildableName = "DatePickerDialogExample.app"
+               BlueprintName = "DatePickerDialogExample"
+               ReferencedContainer = "container:DatePickerDialog.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -29,6 +29,15 @@
       shouldUseLaunchSchemeArgsEnv = "YES">
       <Testables>
       </Testables>
+      <MacroExpansion>
+         <BuildableReference
+            BuildableIdentifier = "primary"
+            BlueprintIdentifier = "4B1E9F361D8F3C9A00C1EF25"
+            BuildableName = "DatePickerDialogExample.app"
+            BlueprintName = "DatePickerDialogExample"
+            ReferencedContainer = "container:DatePickerDialog.xcodeproj">
+         </BuildableReference>
+      </MacroExpansion>
       <AdditionalOptions>
       </AdditionalOptions>
    </TestAction>
@@ -42,15 +51,16 @@
       debugDocumentVersioning = "YES"
       debugServiceExtension = "internal"
       allowLocationSimulation = "YES">
-      <MacroExpansion>
+      <BuildableProductRunnable
+         runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "BC24AA5A1CDB5D7A006438C4"
-            BuildableName = "DatePickerDialogFramework.framework"
-            BlueprintName = "DatePickerDialogFramework"
-            ReferencedContainer = "container:DatePickerDialogExample.xcodeproj">
+            BlueprintIdentifier = "4B1E9F361D8F3C9A00C1EF25"
+            BuildableName = "DatePickerDialogExample.app"
+            BlueprintName = "DatePickerDialogExample"
+            ReferencedContainer = "container:DatePickerDialog.xcodeproj">
          </BuildableReference>
-      </MacroExpansion>
+      </BuildableProductRunnable>
       <AdditionalOptions>
       </AdditionalOptions>
    </LaunchAction>
@@ -60,15 +70,16 @@
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
       debugDocumentVersioning = "YES">
-      <MacroExpansion>
+      <BuildableProductRunnable
+         runnableDebuggingMode = "0">
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "BC24AA5A1CDB5D7A006438C4"
-            BuildableName = "DatePickerDialogFramework.framework"
-            BlueprintName = "DatePickerDialogFramework"
-            ReferencedContainer = "container:DatePickerDialogExample.xcodeproj">
+            BlueprintIdentifier = "4B1E9F361D8F3C9A00C1EF25"
+            BuildableName = "DatePickerDialogExample.app"
+            BlueprintName = "DatePickerDialogExample"
+            ReferencedContainer = "container:DatePickerDialog.xcodeproj">
          </BuildableReference>
-      </MacroExpansion>
+      </BuildableProductRunnable>
    </ProfileAction>
    <AnalyzeAction
       buildConfiguration = "Debug">
