+++ Demo/AppDelegate.swift
@@ -1,0 +1,6 @@
+import UIKit
+
+@UIApplicationMain
+class AppDelegate: UIResponder, UIApplicationDelegate {
+    var window: UIWindow?
+}
