--- Demo/DatePickerDialogExample.xcodeproj/project.xcworkspace/xcshareddata/DatePickerDialogExample.xccheckout
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>IDESourceControlProjectFavoriteDictionaryKey</key>
-	<false/>
-	<key>IDESourceControlProjectIdentifier</key>
-	<string>CB7F4BF7-53BC-46C6-9434-BCE8DF39E2DC</string>
-	<key>IDESourceControlProjectName</key>
-	<string>DatePickerDialogExample</string>
-	<key>IDESourceControlProjectOriginsDictionary</key>
-	<dict>
-		<key>D310D4916D881D443225FAC3D55180619F342255</key>
-		<string>https://github.com/squimer/DatePickerDialog-iOS-Swift.git</string>
-	</dict>
-	<key>IDESourceControlProjectPath</key>
-	<string>Demo/DatePickerDialogExample.xcodeproj</string>
-	<key>IDESourceControlProjectRelativeInstallPathDictionary</key>
-	<dict>
-		<key>D310D4916D881D443225FAC3D55180619F342255</key>
-		<string>../../..</string>
-	</dict>
-	<key>IDESourceControlProjectURL</key>
-	<string>https://github.com/squimer/DatePickerDialog-iOS-Swift.git</string>
-	<key>IDESourceControlProjectVersion</key>
-	<integer>111</integer>
-	<key>IDESourceControlProjectWCCIdentifier</key>
-	<string>D310D4916D881D443225FAC3D55180619F342255</string>
-	<key>IDESourceControlProjectWCConfigurations</key>
-	<array>
-		<dict>
-			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
-			<string>public.vcs.git</string>
-			<key>IDESourceControlWCCIdentifierKey</key>
-			<string>D310D4916D881D443225FAC3D55180619F342255</string>
-			<key>IDESourceControlWCCName</key>
-			<string>DatePickerDialog-iOS-Swift</string>
-		</dict>
-	</array>
-</dict>
-</plist>
